portuguese: translation improvements, better call-to-action
The translation for "Lets-get-arrested project" is more about
"Projeto vamos-ser-presos" in portuguese, similar to
"Proyecto que-nos-arresten" already used in Spanish. There are other small
variations that could be used, but the actual actual translation
"Projeto todos-foram-presos-juntos" in English is more near the meaning
"Project everyone-were-arrested-together".

- Então se isso é considerado um "crime" no Japão. Então vamos ser criminosos e ser presos!
+ Então se isso é considerado um "crime" no Japão. Então vamos ser criminosos e presos!

This change makes make more clear that there is a difference between just be a
criminal and be arrested. The older explanation assumes that only because is a
criminal would be arrested.

- ## Não há ninguem para prendê-lo ?
+ ## Não te prenderam?

This is a minor change. The old one could still work fine (just missed 'é',
woud be "ninguém", not "ninguem") and alone would not be really strong reason
to make a pull request, but since I'm already doing a pull request this is a bit
better.